.revenue method for Vendor class, fixed @amount attr for Sale class to integer, cleaned up debug notes across all classes

+++ lib/far_mar/sale.rb
@@ -1,21 +1,19 @@
 ## TO DO
-# -make sure formatting (.to_i, .to_f, etc.) is correct in initialize, esp for $ (cents)
 # -refactor time formatting w/ .strftime for better readability
 # -refactor self.between method for more user_friendly input (not assign variables, use Time.parse)
-# -commit: Sale class .between method (assuming user assigns beg_ and end_time variables using Time.parse)
+
+##COMMIT NOTES:
+# -
 
 require 'csv'
 
 class FarMar::Sale
 
-  # Each sale belongs to a vendor AND an product, the vendor_id and product_id
-  # fields relates to the Vendor and Product ID fields
-
   attr_accessor :id, :amount, :purchase_time, :vendor_id, :product_id
 
   def initialize (attrs)
     @id             = attrs[:id].to_i
-    @amount         = attrs[:amount].to_f # + /100 to end? needs to be in cents
+    @amount         = attrs[:amount].to_i
     @purchase_time  = Time.parse(attrs[:purchase_time]) # => add .strftime later
     @vendor_id      = attrs[:vendor_id].to_i
     @product_id     = attrs[:product_id].to_i
@@ -26,7 +24,7 @@ class FarMar::Sale
   def self.read_file
     all_sales = CSV.read("/Users/danieladler/Programming/Code_Builders/ruby_practice/projects/far_mar/support/sales.csv", "r")
 
-    attributes = all_sales.map do |s| # will return array of hashes, each containing all info per market
+    attributes = all_sales.map do |s| # ret arr of hashes, each w/ all info per market
       { id:             s[0],
         amount:         s[1],
         purchase_time:  s[2],
@@ -46,7 +44,6 @@ class FarMar::Sale
   end
 
 ### UNIQUE CLASS METHODS:
-
   def vendor # ret Vendor instance assoc w/ this sale via Sale vendor_id
     all_vendors = FarMar::Vendor.all
     match = all_vendors.find {|v| v.id == self.vendor_id}
